requirements.txt
pyrogram>=2.0.30
TgCrypto
uvloop
lexica-api
httpx[http2]
python-dotenv
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
requests
bs4
git+https://github.com/Joelkb/cinemagoer
python-dotenv
Heroku3
search_engine_parser
wget
aiofiles
ffmpeg-python
 ffmpeg
youtube_search
youtube-search-python==1.4.6
yt-dlp
youtube-dl
requests 
aiohttp==3.7.4
gTTS
googletrans==3.1.0a0
telegraph
PyPDF2
ujson
pyshorteners
countryinfo
datetime
pytz
dnspython==2.3.0
aiohttp
python-math
Pillow
speedtest-cli
openai==0.27.0
google-generativeai
